Moved pack_format logic
