--- data/pack_formats.json
@@ -1,10 +1,0 @@
-[
-  {
-    "pack_format": 7,
-    "game_versions": "1.17"
-  },
-  {
-    "pack_format": 8,
-    "game_versions": "1.18+"
-  }
-]
